feat: implement event handler system, fix #98
