+++ teapot/events.py
@@ -9,7 +9,7 @@ def __init__(bot):
     join(bot)
     leave(bot)
     on_guild_join(bot)
-    message_send(bot)
+    # on_message is now handled by event_handler/loader.py
     message_edit(bot)
     message_delete(bot)
     on_command_error(bot)
@@ -50,79 +50,6 @@ def on_guild_join(bot):
     async def on_guild_join(ctx):
         if teapot.config.storage_type() == "mysql":
             teapot.database.create_guild_table(ctx.guild)
-
-
-def message_send(bot):
-    @bot.event
-    async def on_message(message):
-        if teapot.config.storage_type() == "mysql":
-            try:
-                database = teapot.database.__init__()
-                db = teapot.database.db(database)
-                db.execute("SELECT * FROM `users` WHERE user_id = '" + str(message.author.id) + "'")
-                if db.rowcount == 0:
-                    db.execute("INSERT INTO `users`(user_id, user_name, user_display_name) VALUES(%s, %s, %s)",
-                               (message.author.id, message.author.name, message.author.display_name))
-                    database.commit()
-
-                db.execute("SELECT * FROM `channels` WHERE channel_id = '" + str(message.channel.id) + "'")
-                if db.rowcount == 0:
-                    db.execute("INSERT INTO `channels`(channel_id, channel_name) VALUES(%s, %s)",
-                               (message.channel.id, message.channel.name))
-                    database.commit()
-                db.execute(
-                    "INSERT INTO `guild_logs`(timestamp, guild_id, channel_id, message_id, user_id, action_type, message) VALUES(%s, %s, %s, %s, %s, %s, %s)",
-                    (teapot.time(), message.guild.id, message.channel.id, message.id, message.author.id,
-                     "MESSAGE_SEND", message.content))
-                database.commit()
-            except Exception as e:
-                print(e)
-        await bot.process_commands(message)
-
-    @bot.event
-    async def on_message(message):
-        # SAO Easter Egg
-        punctuations = '!()-[]{};:\'"\\,<>./?@#$%^&*_~'
-        # remove punctuation from the string
-        msg = ""
-        for char in message.content.lower():
-            if char not in punctuations:
-                msg = msg + char
-
-        # profanity check
-        prob = predict_prob([msg])
-        if prob >= 0.8:
-            em = discord.Embed(title=f"AI Analysis Results", color=0xC54B4F) # TODO: this will be replaced with cloud detection soon
-            em.add_field(name='PROFANITY DETECTED! ', value=str(prob[0]))
-            await message.channel.send(embed=em)
-
-        if msg.startswith("system call "):
-            content = msg[12:].split(" ")
-            if content[0].lower() == "inspect":
-                if content[1].lower() == "entire":
-                    if content[2].lower() == "command":
-                        if content[3].lower() == "list":
-                            em = discord.Embed(title=f"🍢 SAO Command List", color=0x7400FF)
-                            em.set_thumbnail(
-                                url="https://cdn.discordapp.com/attachments/668816286784159763/674285661510959105/Kirito-Sao-Logo-1506655414__76221.1550241566.png")
-                            em.add_field(name='Commands',
-                                         value="generate xx element\ngenerate xx element xx shape\ninspect entire command list")
-
-                            em.set_footer(text=f"{teapot.copyright()} | Code licensed under the MIT License")
-                            await message.channel.send(embed=em)
-            elif content[0].lower() == "generate":
-                if content[-1].lower() == "element":
-                    em = discord.Embed(title=f"✏ Generated {content[1].lower()} element!",
-                                       color=0xFF0000)
-                    await message.channel.send(embed=em)
-                if content[-1].lower() == "shape":
-                    if content[2].lower() == "element":
-                        em = discord.Embed(
-                            title=f"✏ Generated {content[-2].lower()} shaped {content[1].lower()} element!",
-                            color=0xFF0000)
-                        await message.channel.send(embed=em)
-        await bot.process_commands(message)
-
 
 def message_edit(bot):
     @bot.event
